fix infer bug of NEP GPU

+++ src/feature/NEP_GPU/force/nep_functions.cuh
@@ -559,7 +559,7 @@ static __global__ void find_force_ZBL(
   int n1 = blockIdx.x * blockDim.x + threadIdx.x + N1;
   if (n1 < N) {
     int type1 = g_type[n1];
-    float zi = zbl.atomic_numbers[type1];
+    int zi = zbl.atomic_numbers[type1];
     float pow_zi = pow(zi, 0.23f);
     for (int i1 = 0; i1 < g_NN[n1]; ++i1) {
       int index = n1 + N * i1;
@@ -574,7 +574,7 @@ static __global__ void find_force_ZBL(
       }
       float d12inv = 1.0f / d12;
       float f, fp;
-      float zj = zbl.atomic_numbers[type2];
+      int zj = zbl.atomic_numbers[type2];
       float a_inv = (pow_zi + pow(zj, 0.23f)) * 2.134563f;
       float zizj = K_C_SP * zi * zj;
       if (zbl.flexibled) {
